feat: Add functionality to specify dates for tickets
This adds the possibility to also use the already specified
fields to only download the project time for a specified time
span. The default is to download all tickets according to the
regex specified in the previous commits, by setting the default
for the from_date to 1970-01-01 and the to_date to today's date.
If no ticket-id is found, all matching ticket-ids are used.